Move constants to different file
